Refactor session concurrency tests
