+++ drools-compiler/src/test/java/org/drools/compiler/integrationtests/facts/CategoryTypeEnum.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2017 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.drools.compiler.integrationtests.facts;
+
+import org.drools.compiler.integrationtests.concurrency.AbstractConcurrentTest;
+
+public enum CategoryTypeEnum {
+    ODD, PAIR;
+
+    public static CategoryTypeEnum fromString(final String s) {
+        if (s.equalsIgnoreCase( "odd" )) {
+            return ODD;
+        }
+        if (s.equalsIgnoreCase( "pair" )) {
+            return PAIR;
+        }
+        throw new RuntimeException( "Unknown category: " + s );
+    }
+}
